- Enable 48-bit LBA support in hypervisor.

+++ cube/patches/base/blockdevice.c
@@ -34,7 +34,7 @@ static struct {
 
 OSAlarm read_alarm = {0};
 
-bool do_read_disc(void *buffer, uint32_t length, uint32_t offset, uint32_t sector, frag_callback callback)
+bool do_read_disc(void *buffer, uint32_t length, uint32_t offset, uint64_t sector, frag_callback callback)
 {
 	return do_read_write_async(buffer, length, offset, sector, false, callback);
 }
