+++ cube/patches/dvd/dvd.c
@@ -87,7 +87,7 @@ static void gcode_set_disc_number(uint32_t disc)
 }
 #endif
 
-bool do_read_disc(void *buffer, uint32_t length, uint32_t offset, uint32_t sector, frag_callback callback)
+bool do_read_disc(void *buffer, uint32_t length, uint32_t offset, uint64_t sector, frag_callback callback)
 {
 	return false;
 }
